requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
celery==4.4.2
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
mysqlclient==2.1.1
redis==3.5.3
requests==2.28.1